feat: 신규 랜딩페이지를 메인(/)으로 통합 및 기존 파일 정리
